Remove duplicate tests after merging. Moving new tests under the redteam folder of tests.
