+++ applications/redeye-e2e/src/integration/e2e/redteam/command-count.cy.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
 
-import { graphqlRequest } from '../../support/utils';
+import { graphqlRequest } from '../../../support/utils';
 
 describe('Command counts', () => {
 	const camp = 'commandcounts';
