+++ applications/redeye-e2e/src/integration/e2e/redteam/command-to-beacon.cy.js
