+++ applications/redeye-e2e/src/integration/e2e/redteam/expand-collapse-explorer-panel.cy.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
 
-import { graphqlRequest } from '../../support/utils';
+import { graphqlRequest } from '../../../support/utils';
 
 describe('Expand/collapse explorer panel', () => {
 	const camp = 'expandcollapse';
