+++ applications/redeye-e2e/src/integration/e2e/redteam/logout.cy.js
