+++ applications/redeye-e2e/src/integration/e2e/redteam/search-result-count.cy.js
